+++ rosbag2_transport/src/rosbag2_transport/config_options_from_node_params.cpp
@@ -425,7 +425,7 @@ RecordOptions get_record_options_from_node_params(rclcpp::Node & node)
             "record.repeat_all_transient_local depth must be a non-negative integer.");
   }
   record_options.repeat_all_transient_local_depth =
-    static_cast<size_t>(repeat_all_transient_local_depth);
+    static_cast<uint32_t>(repeat_all_transient_local_depth);
 
   record_options.use_sim_time = node.get_parameter("use_sim_time").get_value<bool>();
